requirements.txt
pandas
numpy
matplotlib
seaborn
openpyxl
python-dotenv
fyers-apiv3
requests
pytz